Move arg parsing to static utility class
