+++ CliToolkit/Utilities/ICommand.cs
@@ -1,0 +1,7 @@
+namespace CliToolkit.Utilities
+{
+    public interface ICommand
+    {
+         void OnExecute(string[] args);
+    }
+}
